feat: Moved message handling to a function
Note: It's really hard to get a Rust panic which is completely
blank. I think you can do it if you use unsafe code, but
I don't see a reason to test for that..? Especially if we're
currently telling users that we couldn't find a solid reason.

In the future, if `None` panics become more common, I'll consider
adjusting the behavior here. No reason at this time! :)